pyproject.toml
[project]
name = "llm-svg-generator-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "openai>=2.8.1",
    "ruff>=0.14.6",
    "uvicorn>=0.38.0",
]
